various edits, including renaming ElectronScattering integral, interactve mode for elspectro, ...

+++ core/src/elspectro.cpp
@@ -10,19 +10,25 @@ int main(int argc, char **argv) {
 
   //get command line options first check if makeall
   TString macroName;
-  for(Int_t i=1;i<argc;i++){
+  bool isInteractive=false;
+  for(Int_t i=0;i<argc;i++){
     TString opt=argv[i];
     if((opt.Contains(".C"))) macroName=opt;
+    else if(opt==TString("--i")) isInteractive=true;
   }
   
   TRint  *app = new TRint("elSpectro", &argc, argv);
   // Run the TApplication (not needed if you only want to store the histograms.)
   app->ProcessLine(".x $ELSPECTRO/core/src/Load.C");
 
-  app->ProcessLine(Form(".x %s",macroName.Data()));
- 
 
-  app->Terminate(0);
+
+  if(isInteractive == true)
+    app->Run();
+  else{
+   app->ProcessLine(Form(".x %s",macroName.Data()));
+   app->Terminate(0);
+  }
   
   return 0;
